+++ src/frontend/js/hooks/useOrders.ts
@@ -1,9 +1,9 @@
 import { useQueryClient } from 'react-query';
 import { useJoanieApi } from 'data/JoanieApiProvider';
-import useLocalizedQueryKey from 'utils/react-query/useLocalizedQueryKey';
 import { REACT_QUERY_SETTINGS } from 'settings';
-import { useSessionQuery } from 'utils/react-query/useSessionQuery';
+import useLocalizedQueryKey from 'utils/react-query/useLocalizedQueryKey';
 import { useSessionMutation } from 'utils/react-query/useSessionMutation';
+import { useSessionQuery } from 'utils/react-query/useSessionQuery';
 
 /**
  * Joanie Api hook to retrieve/create/abort an order owned by the authenticated user.
@@ -28,12 +28,7 @@ export const useOrders = () => {
   };
 
   const creationHandler = useSessionMutation(API.user.orders.create, {
-    onSuccess: async (order) => {
-      // Invalidate orders query
-      await invalidate();
-      // Invalidate course's query related to the creatd order.
-      await queryClient.invalidateQueries(['user', 'course', order.course]);
-    },
+    onSuccess: invalidate,
   });
 
   const abortHandler = useSessionMutation(API.user.orders.abort);